server/requirements.txt
asposestorage==1.0.2
beautifulsoup4==4.14.3
openenv_core==0.2.3
openenv_core.egg==info
